requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pydantic
pydantic-settings
python-dotenv
python-jose
passlib[bcrypt]
bcrypt==4.0.1
cryptography
email-validator
slowapi
python-multipart
httpx>=0.24.0
redis>=5.0.0
google-auth>=2.29.0
boto3>=1.34.0
qrcode[pil]>=7.4.0
